fix: Nova classe criada para auxiliar CreditPayment
