+++ src/utils/CreditPaymentHelpers.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace VindiPaymentGateways;
+
+class CreditHelpers
+{
+
+    public function get_cart_total($cart)
+    {
+        $total = $cart->total;
+        $recurring = end($cart->recurring_carts);
+        if (floatval($cart->total) == 0 && is_object($recurring)) {
+            $total = $recurring->total;
+        }
+        foreach ($cart->get_fees() as $index => $fee) {
+            if ($fee->name == __('Juros', VINDI)) {
+                $total -= $fee->amount;
+            }
+        }
+        return $total;
+    }
+
+}
